Create Maximum Number of Vowels in a Substring of Given Length(o(n^2)).cpp
he time complexity of the given code is O(n^2), where n is the length of the input string s. This is because there are two nested loops, each iterating up to n times.

The space complexity is O(1), as the space used by the code does not depend on the size of the input string or any other input parameters. The only variables declared are count, m, n, i, j, and k, each of which is of constant size. Therefore, the space used by the code is constant and does not depend on the input size.