Create union of two arrays.cpp
Time Complexity: O(n)
Space Complexity: o(n)
...
The space complexity of this function is O(n+m), where n and m are the sizes of the input arrays a[] and b[] respectively. This is because the function uses a map to store the frequency of each element in both arrays, and the size of the map will be at most n+m. Therefore, the space complexity is linear with respect to the size of the input arrays.