Create Reverse The array.cpp
The time complexity of this code is O(m/2), which is equivalent to O(m) because the constant factor is dropped in the big O notation. This is because the for loop iterates m/2 times, so the time complexity is proportional to the input size m.
Time Complexity: o(m)

The space complexity of this code is O(1), because the only additional space used is the temporary variable used in the swap function, which has a constant size and does not depend on the input size m.
Space Complexity: o(1)

Therefore, the space complexity is constant, and the time complexity is linear.